requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
requests
streamlit
httpx
python-dotenv
websocket-client
